+++ src/interpreter/plugin/FunctionPlugin.ts
@@ -12,7 +12,13 @@ import {SearchStrategy} from '../../Lookup/SearchStrategy'
 import {Maybe} from '../../Maybe'
 import {Ast, AstNodeType, ProcedureAst} from '../../parser'
 import {Serialization} from '../../Serialization'
-import {coerceScalarToBoolean, coerceScalarToString, coerceToRange, complex} from '../ArithmeticHelper'
+import {
+  coerceRangeToScalar,
+  coerceScalarToBoolean,
+  coerceScalarToString,
+  coerceToRange,
+  complex
+} from '../ArithmeticHelper'
 import {Interpreter} from '../Interpreter'
 import {InterpreterState} from '../InterpreterState'
 import {
@@ -206,7 +212,7 @@ export abstract class FunctionPlugin implements FunctionPluginTypecheck<Function
 
   protected coerceScalarToNumberOrError = (arg: InternalScalarValue): ExtendedNumber | CellError => this.interpreter.arithmeticHelper.coerceScalarToNumberOrError(arg)
 
-  protected coerceToType(arg: InterpreterValue, coercedType: FunctionArgument): Maybe<InterpreterValue | complex | RawNoErrorScalarValue> {
+  protected coerceToType(arg: InterpreterValue, coercedType: FunctionArgument, state: InterpreterState): Maybe<InterpreterValue | complex | RawNoErrorScalarValue> {
     let ret
     if (arg instanceof SimpleRangeValue) {
       switch(coercedType.argumentType) {
@@ -214,12 +220,13 @@ export abstract class FunctionPlugin implements FunctionPluginTypecheck<Function
         case ArgumentTypes.ANY:
           ret = arg
           break
-        default:
-          if(arg.isAdHoc()) {
-            arg = arg.data[0][0]
-          } else {
+        default: {
+          const coerce = coerceRangeToScalar(arg, state)
+          if(coerce === undefined) {
             return undefined
           }
+          arg = coerce
+        }
       }
     }
     if(!(arg instanceof SimpleRangeValue)) {
@@ -339,7 +346,7 @@ export abstract class FunctionPlugin implements FunctionPluginTypecheck<Function
         }
       } else {
         //we apply coerce only to non-default values
-        const coercedArg = val !== undefined ? this.coerceToType(arg, argumentDefinitions[j]) : arg
+        const coercedArg = val !== undefined ? this.coerceToType(arg, argumentDefinitions[j], state) : arg
         if (coercedArg !== undefined) {
           if (coercedArg instanceof CellError && argumentDefinitions[j].argumentType !== ArgumentTypes.SCALAR) {
             //if this is first error encountered, store it