+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added performance improvements for array handling. (#629)
 - Added ARRAYFORMULA function. (#630)
 - Added casting to scalars from non-range arrays. (#663)
+- Added support for range interpolation. (#TODO)
 
 ### Fixed
 - Fixed an issue with arrays and cruds. (#651)